Move code and add executable
