+++ CMakeLists.txt
@@ -23,3 +23,5 @@ if (NOT NO_GFX)
 else()
 	message("NO_GFX is ON. Base Station is excluded from this build.")
 endif()
+
+add_subdirectory(src/subsystem)
