+++ src/subsystem/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(subsystem subsystem.cpp drive_controller.hpp drive_controller.cpp)
