+++ src/subsystem/drive_controller.cpp
@@ -1,0 +1,2 @@
+#include "drive_controller.hpp"
+// Separate the implementation from drive_controller.hpp and place here
